scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["jukvalim <jukka.valimaa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.9.3"
Scrapy = "^2.4.0"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
